line parser added

+++ atm-department/src/main/java/ru/otus/internals/ATM.java
@@ -4,7 +4,7 @@ import ru.otus.currency.Banknote;
 import ru.otus.currency.Currency;
 
 public interface ATM {
-    void addBanknote(Banknote note, long count);
+    void addBanknote(Banknote note, int count);
     void getCash(long value, Currency currency) throws RuntimeException;
     long getBalance();
 }
