+++ atm-department/src/main/java/ru/otus/internals/ATMBuilder.java
@@ -12,7 +12,7 @@ public class ATMBuilder {
         object = new Bankautomat();
         return object;
     }
-    public ATMBuilder addBanknote(Banknote note, long count) {
+    public ATMBuilder addBanknote(Banknote note, int count) {
 
         getObject().addBanknote(note, count);
         return this;
